// src/app/models/admission-list.model.ts
export class AdmissionList {

    arrivalDate: Date;
    simpleArrivalDate: String;
    simpleArrivalHour: String;
    basicConfirmation: boolean;
    site: boolean;
    anesthesiaSecurity: boolean;
    pulsometer: boolean;
    allergy: boolean;
    difficultAirway: boolean;
    intervention: Array<string>;
    interventionDate: Date;
    simpleInterventionDate: String;
    simpleInterventionHour: String;
    checkDate: Date;
    simpleCheckDate: String;
    simpleCheckHour: String;
    riskOfHemorrhage: boolean;
    status: string;
  
  }
  